+++ script/post-install/index.ts
@@ -1,0 +1,13 @@
+import { npmInstallTip } from 't-comm/lib/npm-tip/npm-tip';
+
+npmInstallTip({
+  packageName: 'uni-plugin-light',
+  packageLink: 'https://novlan1.github.io/uni-plugin-light/',
+  packagePostfix: '工具',
+  packagePostfixEn: 'tools',
+  feedbackList: [
+    'https://github.com/novlan1/uni-plugin-light/issues',
+    'https://github.com/novlan1',
+  ],
+});
+