+++ src/plugin/gen-version/web.ts
@@ -1,31 +1,13 @@
-import { getVersionCode, getCommitCode } from './helper';
+import { getWebInsertCode } from './helper';
 import { updateAssetSource } from '../../helper';
 
 export class GenVersionWebPlugin {
   options: object;
-  buildName?: string;
-  commitName?: string;
-  delay: number;
 
   constructor(options: Record<string, any> = {}) {
     this.options = options || {};
-    this.buildName = options.buildName || '';
-    this.commitName = options.commitName || '';
-    this.delay = options.delay === undefined ? 10 : options.delay;
   }
 
-  getInsertCode() {
-    return `
-<script>
-try {
-  setTimeout(() => {   
-    ${getVersionCode(this.buildName)}
-    ${getCommitCode(this.commitName)}
-  }, ${this.delay});
-} catch(err) {}
-</script>
-`;
-  }
 
   apply(compiler) {
     compiler.hooks.emit.tap('genVersionPlugin', (compilation) => {
@@ -35,7 +17,7 @@ try {
         if (!assets[key]) return;
 
         const source = assets[key].source().toString();
-        const insertCode = this.getInsertCode();
+        const insertCode = getWebInsertCode(this.options);
 
         const idx = source.lastIndexOf('</body>');
         const newSource = source.slice(0, idx) + insertCode + source.slice(idx);