+++ src/vite-plugin/gen-version/index.ts
@@ -1,0 +1,2 @@
+export { genVersionMpVitePlugin } from './mp';
+export { genVersionWebVitePlugin } from './web';