+++ src/vite-plugin/index.ts
@@ -1,0 +1,3 @@
+export * from './gen-version';
+export * from './rem-to-rpx';
+export * from './transform-v-lazy';