+++ src/vite-plugin/rem-to-rpx/index.ts
@@ -1,0 +1,1 @@
+export { remToRpxVitePlugin } from './rem-to-rpx';