+++ src/vite-plugin/transform-v-lazy/index.ts
@@ -1,0 +1,1 @@
+export { transformVLazyVitePlugin } from './transform-v-lazy';