+++ src/vite-plugin/transform-v-lazy/transform-v-lazy.ts
@@ -1,0 +1,14 @@
+import { vLazyCore } from '../../loader/v-lazy/core';
+
+export function transformVLazyVitePlugin() {
+  return {
+    name: 'transform-v-lazy',
+    enforce: 'pre',
+    transform(source: string) {
+      return {
+        code: vLazyCore(source),
+        map: null,
+      };
+    },
+  };
+}